requirements.txt
Django==3.1.1
django-secrets==1.0.2
psycopg2-binary==2.8.6
django_extensions==3.0.9
pyjwt
pyjwe
requests
gunicorn
whitenoise
